Generate null data by simulating from a time series model.

Null hypothesis: data follows a time series model using auto.arima from the forecast package

null_ts(var, modelfn)

Arguments

var
variable to model as a time series

modelfn
method for simulating from ts model.

...
other arguments passed onto modelfn.

Value

a function that given data generates a null data set. For use with lineup or rorschach

See also

Examples

require(forecast)

#> Loading required package: forecast

require(tidyverse)

#> Loading required package: tidyverse

#> ── Attaching packages ─────────────── tidyverse 1.2.1 ──

#> ✔ ggplot2 3.0.0     ✔ readr   1.1.1
#> ✔ tibble  1.4.2     ✔ purrr   0.2.5
#> ✔ tidyr   0.8.1     ✔ stringr 1.3.1
#> ✔ ggplot2 3.0.0     ✔ forcats 0.3.0

#> ── Conflicts ────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
#> ✖ dplyr::n()      masks nullabor::n()

data(aud)

#> Warning: data set ‘aud’ not found

l <- lineup(null_ts("rate", auto.arima), aud)

#> Error in find_plot_data(true): object 'aud' not found

ggplot(l, aes(x=date, y=rate)) + geom_line() +
  facet_wrap(~.sample, scales="free_y") +
  theme(axis.text = element_blank()) +
  xlab("") + ylab("")

#> Error in ggplot(l, aes(x = date, y = rate)): object 'l' not found

l_dif <- l %>%
  group_by(.sample) %>%
  mutate(d=c(NA,diff(rate))) %>%
  ggplot(aes(x=d)) + geom_density() +
  facet_wrap(~.sample)

#> Error in eval(lhs, parent, parent): object 'l' not found