Power calculations.

This function simply counts the proportion of people who selected the data plot, in a set of lineups. It adjusts for multiple picks by the same individual, by weighting by the total number of choices.

visual_power(data, m = 20)

Arguments

data

summary of the results, containing columns id, pic_id, response, detected

m

size of the lineup

Value

vector of powers for each pic_id

Examples

data(turk_results)
visual_power(turk_results)
#> # A tibble: 6 × 3
#>   pic_id power     n
#>    <int> <dbl> <int>
#> 1     36 0        18
#> 2    105 0.746    17
#> 3    116 0.125    16
#> 4    131 0.842    14
#> 5    159 0.656    15
#> 6    225 0.130    15